SPENNY-main-3/SPENNY-main/requirements.txt
python-dotenv==1.0.0
pandas>=2.0.0
numpy>=1.24.0
pandas_ta>=0.3.14b
matplotlib>=3.7.0
pytz>=2023.3
websockets==10.4
aiohttp==3.8.2
alpaca-trade-api==3.1.1
fastapi>=0.100.0
uvicorn>=0.23.0
pyautogui>=0.9.54
pytesseract>=0.3.10
opencv-python>=4.8.0
python-dateutil>=2.8.2
